+++ NEWS.md
@@ -5,6 +5,10 @@
   truncation points - this makes random generation more
   time-efficient
 * Corrected and simplified documentation for `*prop` distribution
+* categorical, multinomial, mixture of normals and mixture of Poisson
+  distributions are now *less restrictive* about probability parameters
+  and accept any non-negtive values. Probability parameter vectors are
+  normalized to sum to one (i.e `c(1,1,1)` becomes `c(1,1,1)/3`).
 
 ### 1.8.1